Create BB84
In this work, the BB84 quantum cryptography protocol is optimized. The impact of variations in intruders, measured states, data to compare, and their respective theoretical predictions focused on the probabilities obtained for each specific case is also evaluated. It is demonstrated that with 50 states to measure and one intruder, both theoretical and experimental probabilities of not detecting it are sufficiently low, thus ensuring the security of the protocol. Additionally, the use of the Quiskit library is explored as a possibility for optimizing the created code.